Move and unify code
