+++ packages/frontend-core/src/utils/index.js
@@ -10,4 +10,4 @@ export { createWebsocket } from "./websocket"
 export * from "./download"
 export * from "./theme"
 export * from "./settings"
-export * from "./schema"
+export * from "./relatedColumns"
